SCP-3509: add runtime checks for allowed builtins
We need to do some runtime checks based on the protocol version in order
to ban builtins that we support but which shouldn't be allowed yet.
We're going to need to do this somewhat systematically, so this provides
the bones of such an approach.

That is:
- The Plutus API functions get passed a `ProtocolVersion`
- There is a big listing of what builtin functions were introduced in
what protocol version.
- We run a check before running a script that it has no builtins that
shouldn't work.

This introduces a pass over the AST, though, which is potentially
expensive overhead. I'll benchmark to see how bad, we might want to
consider trying to fuse this into deserialization somehow.